Thread for synthesizing sound.
Each thread runs its own audioTrack.

For now, audiotrack is paused by simply not writing into it when a boolean is false.
Probably need to test this if it causes problems, but so far it hasn't been problematic.

Also, currently some stuff is hardcodey, I will make it less hardcodey, but I want to see how it runs for you guys first.